packages/server/pyproject.toml
[project]
name = "nautilus-automatron-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyarrow>=18.0.0",
    "pandas>=2.2.0",
    "pydantic-settings>=2.6.0",
    "nautilus_trader",
]
